Create useDiscount under custom hook folder
refactor: Create useDiscount under custom hook folder

- Create useDiscount under custom hook folder

- Delete useDiscount in customHooks
